Rename tests as FIPSEnabled
