+++ src/test/resources/hudson/plugins/active_directory/ActiveDirectoryDomainFIPSEnabledIntegrationTest/config.xml
